+++ naga/tests/snapshots.rs
@@ -789,6 +789,14 @@ fn convert_wgsl() {
             Targets::SPIRV | Targets::METAL | Targets::HLSL | Targets::WGSL,
         ),
         (
+            "atomicTexture",
+            Targets::SPIRV | Targets::METAL | Targets::GLSL | Targets::HLSL | Targets::WGSL,
+        ),
+        (
+            "atomicOps-float32",
+            Targets::SPIRV | Targets::METAL | Targets::WGSL,
+        ),
+        (
             "atomicCompareExchange-int64",
             Targets::SPIRV | Targets::WGSL,
         ),