+++ wgpu-types/src/counters.rs
@@ -1,6 +1,7 @@
+use alloc::{string::String, vec::Vec};
 #[cfg(feature = "counters")]
-use std::sync::atomic::{AtomicIsize, Ordering};
-use std::{fmt, ops::Range};
+use core::sync::atomic::{AtomicIsize, Ordering};
+use core::{fmt, ops::Range};
 
 /// An internal counter for debugging purposes
 ///
@@ -95,8 +96,8 @@ impl Default for InternalCounter {
     }
 }
 
-impl std::fmt::Debug for InternalCounter {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+impl core::fmt::Debug for InternalCounter {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
         self.read().fmt(f)
     }
 }
@@ -192,7 +193,7 @@ impl fmt::Debug for AllocationReport {
 impl fmt::Debug for AllocatorReport {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         let mut allocations = self.allocations.clone();
-        allocations.sort_by_key(|alloc| std::cmp::Reverse(alloc.size));
+        allocations.sort_by_key(|alloc| core::cmp::Reverse(alloc.size));
 
         let max_num_allocations_to_print = f.precision().unwrap_or(usize::MAX);
         allocations.truncate(max_num_allocations_to_print);
@@ -200,7 +201,7 @@ impl fmt::Debug for AllocatorReport {
         f.debug_struct("AllocatorReport")
             .field(
                 "summary",
-                &std::format_args!(
+                &core::format_args!(
                     "{} / {}",
                     FmtBytes(self.total_allocated_bytes),
                     FmtBytes(self.total_reserved_bytes)