+++ wgpu/src/dispatch.rs
@@ -100,6 +100,9 @@ pub trait InstanceInterface: CommonTraits {
     ) -> Pin<Box<dyn RequestAdapterFuture>>;
 
     fn poll_all_devices(&self, force_wait: bool) -> bool;
+
+    #[cfg(feature = "wgsl")]
+    fn wgsl_language_features(&self) -> crate::WgslLanguageFeatures;
 }
 
 pub trait AdapterInterface: CommonTraits {
@@ -572,7 +575,7 @@ macro_rules! dispatch_types_inner {
         impl $name {
             #[cfg(wgpu_core)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_core(&self) -> &<$wgpu_core_context as InterfaceTypes>::$subtype {
                 match self {
                     Self::Core(value) => value,
@@ -582,7 +585,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(wgpu_core)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_core_opt(&self) -> Option<&<$wgpu_core_context as InterfaceTypes>::$subtype> {
                 match self {
                     Self::Core(value) => Some(value),
@@ -592,7 +595,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(webgpu)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_webgpu(&self) -> &<$webgpu_context as InterfaceTypes>::$subtype {
                 match self {
                     Self::WebGPU(value) => value,
@@ -602,7 +605,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(webgpu)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_webgpu_opt(&self) -> Option<&<$webgpu_context as InterfaceTypes>::$subtype> {
                 match self {
                     Self::WebGPU(value) => Some(value),
@@ -657,7 +660,7 @@ macro_rules! dispatch_types_inner {
         impl $name {
             #[cfg(wgpu_core)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_core(&self) -> &<$wgpu_core_context as InterfaceTypes>::$subtype {
                 match self {
                     Self::Core(value) => value,
@@ -667,7 +670,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(wgpu_core)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_core_mut(&mut self) -> &mut <$wgpu_core_context as InterfaceTypes>::$subtype {
                 match self {
                     Self::Core(value) => value,
@@ -677,7 +680,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(wgpu_core)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_core_opt(&self) -> Option<&<$wgpu_core_context as InterfaceTypes>::$subtype> {
                 match self {
                     Self::Core(value) => Some(value),
@@ -687,7 +690,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(wgpu_core)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_core_mut_opt(
                 &mut self,
             ) -> Option<&mut <$wgpu_core_context as InterfaceTypes>::$subtype> {
@@ -699,7 +702,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(webgpu)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_webgpu(&self) -> &<$webgpu_context as InterfaceTypes>::$subtype {
                 match self {
                     Self::WebGPU(value) => value,
@@ -709,7 +712,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(webgpu)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_webgpu_mut(&mut self) -> &mut <$webgpu_context as InterfaceTypes>::$subtype {
                 match self {
                     Self::WebGPU(value) => value,
@@ -719,7 +722,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(webgpu)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_webgpu_opt(&self) -> Option<&<$webgpu_context as InterfaceTypes>::$subtype> {
                 match self {
                     Self::WebGPU(value) => Some(value),
@@ -729,7 +732,7 @@ macro_rules! dispatch_types_inner {
 
             #[cfg(webgpu)]
             #[inline]
-            #[allow(unused)]
+            #[allow(clippy::allow_attributes, unused)]
             pub fn as_webgpu_mut_opt(
                 &mut self,
             ) -> Option<&mut <$webgpu_context as InterfaceTypes>::$subtype> {